requirements.txt
black
celery==5.3.6
crispy-bootstrap5==2024.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-cid==2.3
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-environ==0.11.2
flake8==7.0.0
freezegun==1.4.0
gunicorn==21.2.0
isort
pre-commit==3.7.0
pre-commit-hooks==4.6.0
psycopg2-binary==2.9.9
pytest==8.1.1
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-xdist==3.5.0
pyupgrade==3.15.2
secure==0.3.0
